+++ TShockAPI/ConfigFile.cs
@@ -443,10 +443,6 @@ namespace TShockAPI
 		[Description("#.#.# = Red/Blue/Green - RGB Colors for broadcasts. Max value: 255.")]
 		public int[] BroadcastRGB = { 127, 255, 212 };
 
-		// TODO: Get rid of this when the old REST permission model is removed.
-		[Description("Whether the REST API should use the new permission model. Note: The old permission model will become depracted in the future.")]
-		public bool RestUseNewPermissionModel = true;
-
 		/// <summary>ApplicationRestTokens - A dictionary of REST tokens that external applications may use to make queries to your server.</summary>
 		[Description("A dictionary of REST tokens that external applications may use to make queries to your server.")]
 		public Dictionary<string, SecureRest.TokenData> ApplicationRestTokens = new Dictionary<string, SecureRest.TokenData>();